📦 [Backup] Synced DRYRUNmain + env.example @ 2025-10-16_02-54-30

+++ utils/ta_engine.py
@@ -92,4 +92,5 @@ def calc_vol_spike(highs, lows, period=14):
         return 1.0
 
     spike = avg_recent / avg_prev
-    return round(spike, 2)
+    return float(round(spike, 2))
+
